Moved logic for save compatibility to a dedicated worker class.
